requirements.txt
omegaconf
accelerate
diffusers==0.20.2
transformers
safetensors
xformers
einops
pillow
torch>=1.13.1
torchvision
tensorboard
torchmetrics
triton
numba
tqdm
ftfy
regex
timm
numpy
scikit-learn
scikit-fmm
scipy
scikit-image
matplotlib
visdom
wandb
opencv-python
BeautifulSoup4
freetype-py
shapely
libs